feat: 1. Cactus 프린트 형식 및 프린트 기능 구현

+++ app/src/main/java/com/cjwgit/jejucactusreceipt/domain/CactusAuctionEntity.kt
@@ -4,6 +4,6 @@ data class CactusAuctionEntity(
     val uid: Long,
     val name: String,
     // 한 박스에 선인장 몇개 있는지
-    val amount: Int,
-    val price: Int,
+    val amount: Long,
+    val price: Long,
 )
